fix(protocol): name_id u16→u8 on wire + ROADMAP v1.4 documentation
name_id wire serialization reduced from 2 bytes to 1 byte per channel
in multi-channel frames. The Rust field type ChannelInput.name_id is
now u8 (max 255 channels per device). Saves 1B × N channels per frame.

Combined with header v2 changes (sequence u16, context_version u24),
total saving vs 1.3.0: up to 8B on a 5-channel payload.

Also adds ROADMAP.md section [v1.4] documenting planned Pattern (0x20)
and Interpolated (0x31) encoding modes (documentation only).

Includes 3 regression tests: test_name_id_1_byte_in_frame,
test_name_id_max_u8, test_name_id_roundtrip.

No version bump — 1.3.1 will be tagged after full validation.

+++ alec-ffi/src/lib.rs
@@ -396,7 +396,7 @@ pub extern "C" fn alec_encode_multi(
             };
 
             ChannelInput {
-                name_id: i as u16,
+                name_id: i as u8,
                 source_id: sid,
                 value: values_slice[i],
             }
